build_requirements.txt

click==7.1.2
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
importlib-metadata==2.0.0
importlib-resources==3.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
paho-mqtt==1.5.1
pyaml==18.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
pygame==1.9.6
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pywebview==3.3.5
PyYAML==5.3.1
requests>=2.25.0
six==1.15.0
Werkzeug==1.0.1
zipp==3.3.1
zope.event==4.5.0
zope.interface==5.1.2
pytest
pytest-flask

